Move ResourcesExtractor to the public interface
