+++ src/libappimage/desktop_integration/CMakeLists.txt
@@ -4,8 +4,6 @@ set(
     integrator/Integrator.cpp
     integrator/DesktopEntryEditError.h
     integrator/DesktopEntryEditor.cpp
-    integrator/ResourcesExtractor.cpp
-    integrator/PayloadEntriesCache.cpp
 )
 
 if(LIBAPPIMAGE_THUMBNAILER_ENABLED)
