+++ src/libappimage/utils/CMakeLists.txt
@@ -7,6 +7,8 @@ add_library(
     IconHandle.cpp
     Logger.cpp
     path_utils.cpp
+    resources_extractor/ResourcesExtractor.cpp
+    resources_extractor/PayloadEntriesCache.cpp
 )
 
 # Generic module configuration
@@ -14,6 +16,7 @@ configure_libappimage_module(appimage_utils)
 target_include_directories(
     appimage_utils
     PRIVATE $<TARGET_PROPERTY:Boost::filesystem,INTERFACE_INCLUDE_DIRECTORIES>
+    PRIVATE $<TARGET_PROPERTY:XdgUtils,INTERFACE_INCLUDE_DIRECTORIES>
     PUBLIC $<TARGET_PROPERTY:libappimage_hashlib,INTERFACE_INCLUDE_DIRECTORIES>
 )
 add_dependencies(appimage_utils Boost::filesystem libappimage_hashlib)
