+++ tests/libappimage/CMakeLists.txt
@@ -15,9 +15,12 @@ add_executable(
     utils/TestMagicBytesChecker.cpp
     utils/TestUtilsElf.cpp
     utils/TestIconHandle.cpp
-    utils/TestLogger.cpp)
+    utils/TestLogger.cpp
+    utils/TestPayloadEntriesCache.cpp
+    utils/TestResourcesExtractor.cpp
+)
 
 target_include_directories(test_libappimage++ PRIVATE "${PROJECT_SOURCE_DIR}/src/libappimage")
-target_link_libraries(test_libappimage++ libappimage Boost::filesystem libarchive libsquashfuse gtest gtest_main)
+target_link_libraries(test_libappimage++ libappimage Boost::filesystem libarchive libsquashfuse XdgUtils gtest gtest_main)
 
 add_test(test_libappimage++ test_libappimage++)
