+++ tests/libappimage/desktop_integration/CMakeLists.txt
@@ -5,8 +5,6 @@ set(
 
     integrator/TestDesktopIntegration.cpp
     integrator/TestDesktopEntryEditor.cpp
-    integrator/TestResourcesExtractor.cpp
-    integrator/TestPayloadEntriesCache.cpp
 
     $<TARGET_OBJECTS:core>
     $<TARGET_OBJECTS:appimage_utils>
